+++ src/views/payment-sheet-views/base-paypal-view.js
@@ -72,14 +72,22 @@ BasePayPalView.prototype.initialize = function () {
       self.paypalConfiguration.locale = locale;
       checkoutJSConfiguration.locale = locale;
     }
+    checkoutJSConfiguration.funding = {
+      disallowed: []
+    };
+
+    Object.keys(global.paypal.FUNDING).forEach(function (key) {
+      if (key === 'PAYPAL' || key === 'CREDIT') {
+        return;
+      }
+      checkoutJSConfiguration.funding.disallowed.push(global.paypal.FUNDING[key]);
+    });
 
     if (isCredit) {
       buttonSelector = '[data-braintree-id="paypal-credit-button"]';
       checkoutJSConfiguration.style.label = 'credit';
     } else {
-      checkoutJSConfiguration.funding = {
-        disallowed: [global.paypal.FUNDING.CREDIT]
-      };
+      checkoutJSConfiguration.funding.disallowed.push(global.paypal.FUNDING.CREDIT);
     }
 
     return global.paypal.Button.render(checkoutJSConfiguration, buttonSelector).then(function () {