+++ CHANGELOG.md
@@ -8,6 +8,7 @@ unreleased
   - Google Pay: Add support for `isNetworkTokenized` param in `parseResponse` method
   - Card Form: Fix issue where pasting a card number over an Amex number could cut off the last digit
   - PayPal: Add support for shipping options (see https://braintree.github.io/braintree-web/current/PayPalCheckout.html#createPayment)
+- Prevent non-PayPal funding sources from appearing in the PayPal views
 
 1.20.4
 ------